Remove uneccessary guard condition from handleClick.
Originally, you could click a card that was already visible and
the handleClick event handler would run, so I had a guard condition
setup to prevent this (but it actually wasn't even fool-proof!). Now,
I'm removing the click handler when a card is shown, so there is
no reason to guard against this.