Removed this.numClicks and refactored how I'm using this.lastClickedCardId.
I had some fuzzy logic going on throughout handleClick() and
checkForMatchedCards() (now called hideUnmatchedCards()). Before,
I was repeating some checks that I didn't catch at first, or
just doing things that weren't logically neccessary, like keeping
track of the number of clicks. I was using lastClickedCardId and
numClicks, but really the lastClickedCardId gives me the same information
as numClicks (the way I'm using it), so numClicks was not needed.
Now, the handleClick method turns over the clicked card, looks for
a match with the lastClickedCard (if there is one), and if so
it is done. If not, it may set lastClickedCard, then it will
possibly run hideUnmathcedCards if the lastClickedCard is not the
same as the currently clicked card. This is still a little confusing,
but getting better.